requirements.txt
# â”€â”€ Core Flask API â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
flask==3.0.3
flask-cors==4.0.0
Flask-Limiter==3.5.0
flask-socketio==5.3.6           # WebSocket/Socket.IO server for real-time metric push
psutil==5.9.8
requests==2.32.3
PyYAML==6.0.1

# â”€â”€ New: FastAPI Auth + Core Services â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic[email]==2.9.2
python-multipart==0.0.17
slowapi==0.1.9              # rate limiting for FastAPI

# â”€â”€ JWT (shared by Flask + FastAPI) â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
passlib[bcrypt]==1.7.4
python3-saml==1.16.0
xmltodict==0.14.2
cryptography==43.0.3

# â”€â”€ PostgreSQL (async for auth_api, sync driver available for Flask) â”€â”€â”€â”€â”€â”€â”€â”€â”€
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10      # sync driver â€” useful for Alembic migrations
alembic==1.13.2               # database migration tool for PostgreSQL

# â”€â”€ AI integrations â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
google-generativeai==0.7.2
openai==1.51.2

# â”€â”€ Async HTTP (notification_service â€” Slack & Telegram dispatch) â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
aiohttp==3.10.5

# â”€â”€ Structured logging (ELK Stack) â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
python-json-logger==2.0.7

# â”€â”€ OpenTelemetry distributed tracing (Jaeger) â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-instrumentation-sqlalchemy==0.48b0
protobuf==4.25.8

# â”€â”€ Optional Firebase (comment out if not using Firebase) â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
# firebase-admin==6.5.0

# â”€â”€ Optional HuggingFace AI â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
# transformers==4.43.3
# torch==2.3.1
# tokenizers==0.19.1
# huggingface-hub==0.23.4

# â”€â”€ Optional Redis (response caching) â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
# redis==5.0.8

# â”€â”€ WMI Agentless Polling (Windows remote metrics via WinRM) â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
pywinrm==0.4.3                 # WinRM client for Python (server polls Windows machines)
# pywinrm[kerberos]            # uncomment for Active Directory / domain auth

# â”€â”€ Dev / Test â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
# pytest==8.2.2
# pytest-asyncio==0.23.7
# httpx==0.27.0   # needed for FastAPI TestClient async

# â”€â”€ Load / Performance Testing â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
locust==2.31.8              # load testing framework â€” run: locust -f tests/load/locustfile.py

